# eee_portal_backend/openapi.yml
openapi: 3.0.3
info:
  title: EEE Portal Backend API
  description: |
    API documentation for the EEE Portal Backend with complete security specifications.
    
    ## Authentication
    - **Bearer Token**: JWT token required for authenticated endpoints
    - **User Roles**: admin, user
    
    ## Security Levels
    - 🔓 Public: No authentication required
    - 🔐 User: Any logged-in user
    - 🔑 Admin: Admin role required
  version: 1.0.0
  contact:
    name: EEE Portal Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.eeportal.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication
  schemas:
...
    Response:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user:
          type: integer
          example: 1
        question:
          type: integer
          example: 1
        options:
          type: integer
          example: 0
        created_at:
          type: string
          format: date-time
...
paths:
...
  /responses/single-response/:
    post:
      summary: Store single response
      description: 🔐 User authentication required - Store single response record for a question
      tags:
        - Responses Service
      operationId: submit_single_response
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
                - options
              properties:
                question:
                  type: integer
                  example: 1
                options:
                  type: integer
                  example: 0
      responses:
        '201':
          description: Response stored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized - Token missing or invalid
  /responses/bulk-response/:
    post:
      summary: Store bulk responses
      description: 🔐 User authentication required - Store bulk response records of same user
      tags:
        - Responses Service
      operationId: submit_bulk_response
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required:
                  - question
                  - options
                properties:
                  question:
                    type: integer
                  options:
                    type: integer
              example:
                - question: 1
                  options: 0
                - question: 2
                  options: 1
                - question: 3
                  options: 2
      responses:
        '201':
          description: Responses stored successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Response'
        '400':
          description: Invalid input or expected list
        '401':
          description: Unauthorized - Token missing or invalid
  /responses/{id}/delete-response/:
    delete:
      summary: Delete user responses
      description: 🔑 Admin only - Delete all response records of user having specified ID
      tags:
        - Responses Service
      operationId: delete_user_response
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: User ID
      responses:
        '200':
          description: Responses deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Deleted 5 records"
        '401':
          description: Unauthorized - Token missing or invalid
        '403':
          description: Forbidden - Admin access required
        '404':
          description: User responses not found
  /responses/{id}/get-response/:
    get:
      summary: Get user responses
      description: 🔑 Admin only - Get all response records of a user with specified ID
      tags:
        - Responses Service
      operationId: get_user_response
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: User ID
      responses:
        '200':
          description: User responses retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Response'
        '401':
          description: Unauthorized - Token missing or invalid
        '403':
          description: Forbidden - Admin access required
        '404':
          description: No responses found for this user
...